Constrain dependencies to current major version
That way we can make backwards incompatible changes in these
dependencies on the next major versions, without breaking this gem.

Even though tus-server isn't strictly required for 2 out of 3 ways to
use this gem, and shrine-url isn't require for the plugin way, we still
include them to be able to control the allowed versions. They will be
loaded only when needed anyway, it's just too risky to trade certainty
for "not having gems in the Gemfile that might not be used".